Move shard indexing tests
